requirements.txt
streamlit
Pillow
google-generativeai
langchain
langchain-google-genai
python-dotenv
